requirements/blogango.txt
Markdown==2.3.1
MarkupSafe==0.23
Pygments==1.6
PyYAML==3.11
django-pingback==0.3
django-taggit==0.12
django-wysiwyg==0.5.1
django-xmlrpc==0.1.3
